Restructured menu
This PR moves the 'Showcases' and 'Frontity CLI' sections to lower down in the menu. This keeps the most important information for the user new to Frontity at the top and in a logical order, i.e.:

* Getting Started
* Learning Frontity
* Deployment

With "ancilliary" information lower down in the menu. I've still tried to keep a reasonably logical ordering here too, with more important information higher up.