Update 'Key Differences' section of the About page
I've made some changes to the last section of this page and suggested updating it to be more general as this is where we talk about some of the advantages that make Frontity different from other React frameworks, not just Gatsby. Please feel free to review it.

I also removed a feature from the previous section that was no longer in the "Frontity Features" page (the link was not working).

I guess it will also be necessary to update the "How Frontity Works" section when the new Architecture page is published.